Code clean-up: frontend
