+++ src/renderer/src/pages/Tags.jsx
@@ -1,29 +1,31 @@
 import * as React from 'react'
 import { useNavigate } from 'react-router-dom'
 import { useAllTags, useCreateTags, useDeleteTag } from '../hooks/tags.js'
-import TagsView from '../views/tags/Tags.jsx'
+import CenterMessage from '../views/app/CenterMessage.jsx'
+import TagsView from '../views/tags/TagsView.jsx'
 
 export default function Tags() {
   const [filterText, setFilterText] = React.useState('')
   const [tagInput, setTagInput] = React.useState('')
 
-  const navigate = useNavigate()
-
-  const dbTags = useAllTags().data || []
-
+  const dbTags = useAllTags()
   const [createTags, isCreating] = useCreateTags()
   const deleteTag = useDeleteTag()
+
+  const navigate = useNavigate()
 
   const handleCreateTags = async (e) => {
     await createTags(tagInput)
     document.getElementById('tagInput').value = ''
   }
 
+  if (dbTags.isLoading) return <CenterMessage msg="Loading..." />
+
   return (
     <TagsView
-      setFilterText={setFilterText}
-      dbTags={dbTags}
       filterText={filterText}
+      setFilterText={setFilterText}
+      dbTags={dbTags.data}
       navigate={navigate}
       handleDeleteTag={deleteTag}
       setTagInput={setTagInput}
