+++ src/renderer/src/views/common/SpinnerOr.jsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import { Spinner } from 'react-bootstrap'
 
-export default function SpinnerOr(props) {
-  return props.isSpinner ? (
+const SpinnerOr = (props) =>
+  props.isSpinner ? (
     <div className="d-flex align-items-center">
       <Spinner className="me-2" />
       <span className="me-2">{props.msg}</span>
@@ -10,4 +10,5 @@ export default function SpinnerOr(props) {
   ) : (
     props.children
   )
-}
+
+export default SpinnerOr
