+++ src/renderer/src/views/galleries/GalleryRow.jsx
@@ -2,29 +2,28 @@ import { Button } from 'react-bootstrap'
 import { Check2, X } from 'react-bootstrap-icons'
 import { Link } from 'react-router-dom'
 
-function GalleryRow({ index, dbGallery, galleryName, deleteGallery }) {
-  return (
-    <tr>
-      <td className="text-end col-1">{index + 1}</td>
-      <td>
-        {dbGallery.isAvailable ? (
-          <Link to={`/gallery/${encodeURIComponent(dbGallery.galleryPath)}`}>{galleryName}</Link>
-        ) : (
-          galleryName
-        )}
-      </td>
-      <td>{dbGallery.isAvailable ? <Check2 /> : <X />}</td>
-      <td>
-        <Button
-          variant="danger"
-          size="sm"
-          onClick={async () => await deleteGallery(dbGallery.galleryPath)}
-        >
-          Delete
-        </Button>
-      </td>
-    </tr>
-  )
-}
+const GalleryRow = ({ dbGallery, deleteGallery }) => (
+  <tr>
+    <td>
+      {dbGallery.isAvailable ? (
+        <Link to={`/gallery/${encodeURIComponent(dbGallery.galleryPath)}`}>
+          {dbGallery.galleryName}
+        </Link>
+      ) : (
+        dbGallery.galleryName
+      )}
+    </td>
+    <td>{dbGallery.isAvailable ? <Check2 /> : <X />}</td>
+    <td>
+      <Button
+        variant="danger"
+        size="sm"
+        onClick={async () => await deleteGallery(dbGallery.galleryPath)}
+      >
+        Delete
+      </Button>
+    </td>
+  </tr>
+)
 
 export default GalleryRow
