+++ src/renderer/src/views/galleries/Operations.jsx
@@ -2,52 +2,52 @@ import * as React from 'react'
 import { Button, Col, Row } from 'react-bootstrap'
 import SpinnerOr from '../common/SpinnerOr.jsx'
 
-export default function Operations({
+const Operations = ({
   isCreating,
   galleryInput,
   getGalleryPathInput,
   handleCreateGallery,
   isDeletingGalleries,
   handleDeleteMissingGalleries
-}) {
-  return (
-    <>
-      <Col xs={12} sm={6}>
-        <Row>
-          <Col>
-            <SpinnerOr isSpinner={isCreating} msg="Creating...">
-              <Button className="me-2" size="sm" onClick={async () => await getGalleryPathInput()}>
-                Add new Gallery
-              </Button>
-              {galleryInput && (
-                <Button
-                  variant="success"
-                  className="ms-2"
-                  size="sm"
-                  onClick={async () => await handleCreateGallery(galleryInput)}
-                >
-                  Submit
-                </Button>
-              )}
-            </SpinnerOr>
-          </Col>
-          <Col>
-            <SpinnerOr isSpinner={isDeletingGalleries} msg="Deleting...">
+}) => (
+  <>
+    <Col xs={12} sm={6}>
+      <Row>
+        <Col>
+          <SpinnerOr isSpinner={isCreating} msg="Creating...">
+            <Button className="me-2" size="sm" onClick={async () => await getGalleryPathInput()}>
+              Add new Gallery
+            </Button>
+            {galleryInput && (
               <Button
-                className="ms-auto"
-                variant="danger"
+                variant="success"
+                className="ms-2"
                 size="sm"
-                onClick={handleDeleteMissingGalleries}
+                onClick={async () => await handleCreateGallery(galleryInput)}
               >
-                Delete Missing Galleries
+                Submit
               </Button>
-            </SpinnerOr>
-          </Col>
-        </Row>
-        <Row>
-          <div>{galleryInput}</div>
-        </Row>
-      </Col>
-    </>
-  )
-}
+            )}
+          </SpinnerOr>
+        </Col>
+        <Col>
+          <SpinnerOr isSpinner={isDeletingGalleries} msg="Deleting...">
+            <Button
+              className="ms-auto"
+              variant="danger"
+              size="sm"
+              onClick={handleDeleteMissingGalleries}
+            >
+              Delete Missing Galleries
+            </Button>
+          </SpinnerOr>
+        </Col>
+      </Row>
+      <Row>
+        <div>{galleryInput}</div>
+      </Row>
+    </Col>
+  </>
+)
+
+export default Operations
